# Witness multiplication optimization